Removed setPretender
